refactor: Moved location of LanguageProvider, removed changeset
