--- .changeset/hip-seahorses-dance.md
@@ -1,14 +1,0 @@
----
-"@navikt/ds-react": major
----
-
-API change for translations
-
-WHAT
-Translations now live on the global provider in addition to allowing per-component overrides.
-
-WHY
-Simplify the translations API, a single point of entry for _all_ components. Work well with i18next.
-
-HOW
-Wrap your whole app with a `<UNSAFE_AkselLanguageProvider translations={{...}}>`.
